templates/003_text_to_sql/pyproject.toml
[project]
name = "text-to-sql"
version = "0.1.0"
description = "Chatbot Template"
readme = "README.md"
authors = [
    { name = "lgcns-ai-keonho-kim", email = "kkh93@lgcns.com" }
]
requires-python = ">=3.13"
dependencies = [
    "elasticsearch>=9.2.1",
    "fastapi>=0.128.0",
    "fastmcp>=2.14.4",
    "httpx>=0.28.1",
    "json-repair>=0.55.1",
    "lancedb>=0.29.2",
    "langchain>=1.2.7",
    "langchain-anthropic>=1.3.1",
    "langchain-google-genai>=4.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "langgraph-checkpoint==3.0.1",
    "langgraph-checkpoint-postgres>=3.0.4",
    "langgraph-checkpoint-redis>=0.3.4",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "lxml>=6.0.2",
    "motor>=3.7.1",
    "olefile>=0.47",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pdfplumber>=0.11.9",
    "pgvector>=0.4.2",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-pptx>=1.0.2",
    "pyyaml>=6.0.3",
    "redis>=7.1.0",
    "redisvl>=0.13.2",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.4",
    "ty>=0.0.18",
]

[project.scripts]
text_to_sql = "text_to_sql:main"

[build-system]
requires = ["uv_build>=0.8.19,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "text_to_sql"
module-root = "src"

[tool.pytest.ini_options]
pythonpath = ["src"]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s | %(levelname)s | %(name)s | %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = "-s -vv"
